switching to thrussh

+++ ssh-forwarding-service/src/interface.rs
@@ -3,13 +3,13 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize)]
 pub struct Input {
     pub ssh_forwarding_config: SshForwardingConfig,
-    pub local_port: u16,
+    pub local_port: u32,
     pub max_polling_retry_times: u16,
 }
 
 #[derive(Serialize, Deserialize)]
 pub struct Output {
-    pub deployed_local_port: u16,
+    pub deployed_local_port: u32,
 }
 
 #[derive(Serialize, Deserialize, Default)]
@@ -18,7 +18,7 @@ pub struct SshForwardingConfig {
     pub ssh_user: String,
     pub ssh_private_key_base64: String,
     pub remote_host: String,
-    pub remote_port: u16,
+    pub remote_port: u32,
 }
 
 impl SshForwardingConfig {
